+++ nltk/model/__init__.py
@@ -6,7 +6,7 @@
 # URL: <http://nltk.org/
 # For license information, see LICENSE.TXT
 
-from nltk.model.api import BaseNgramModel, NEG_INF
+from nltk.model.api import BaseNgramModel
 from nltk.model.models import (MLENgramModel,
                                LidstoneNgramModel,
                                LaplaceNgramModel)
